perf(evm): stub out gas-price metrics in zkevm guest
Split the per-block gas-price aggregates out of Metrics.cs into a partial std
file and add a no-op zkEVM variant. The guest targets rv64ima (no FPU) and
nobody reads the gauges there, so skipping the float CAS loops per transaction
removes proving cost.

(cherry picked from commit ed07bcf0bd6b5025121cf3293bdf4c5c3645defe)
